Make Background Responsive again
Experimentation shows that Svelte's each block is whitespace significant
(where is this in the docs!?), so I've reformatted the markup
appropriately.

Also, have made a separate `chunk` element and reference that to repeat
it, doing it inline would cause a chicken and egg problem where the
thing we want to update itself gets updated when the
`viewPortHeightChanges` or gets updated via `calcRepeats`.